完成用户注册相关，使用一个简单的html页面演示
