+++ src/main/resources/templates/register.html
@@ -1,10 +1,29 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="en"
+      xmlns:th="http://www.thymeleaf.org">
 <head>
     <meta charset="UTF-8">
-    <title>Title</title>
+    <title>XXX——用户注册</title>
 </head>
 <body>
+<div id="content">
 
+    <form method="post" th:action="@{/register}" th:object="${userForm}">
+        <div class="errors" th:if="${#fields.hasErrors()}">
+            <ul>
+                <li th:each="err : ${#fields.errors()}" th:text="${err}">
+                    Input is incorrect;
+                </li>
+            </ul>
+        </div>
+        <label th:class="${#fields.hasErrors('email')}? 'error'">Email : </label>
+        <input th:class="${#fields.hasErrors('email')}? 'error'" th:field="*{email}" type="email"/><br/>
+
+        <label th:class="${#fields.hasErrors('password')}? 'error'">Password : </label>
+        <input th:class="${#fields.hasErrors('password')}? 'error'" th:field="*{password}" type="password"/><br/>
+
+        <input type="submit" value="Register">
+    </form>
+</div>
 </body>
 </html>
